+++ src/common/rabbitmq/consumers/database/database-consumer.service.ts
@@ -7,7 +7,7 @@ import {
 import * as amqp from 'amqplib';
 import { PrismaService } from 'src/modules/prisma/prisma.service';
 import { CreateMessageDto } from './dto/create-message.dto';
-import { MessageValidator } from './Message-validator';
+import { MessageValidator } from './message-validators/Message-validator';
 import { RabbitMQConnectionService } from '../rabbit-connection.service';
 
 @Injectable()
